:sparkles: Init CMS
