+++ Makefile
@@ -1,4 +1,6 @@
 default:
 	yarn
 	cd api && yarn
+	cd data_model && yarn
+	cd cms && yarn
 	cd ui && yarn
