+++ cms/bin/getToken.js
