+++ cms/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "hcmi-portal-api",
-  "version": "0.0.1-0.0.10.0",
-  "description": "``` npm i @arranger/server ```",
+  "name": "hcmi-portal-cms",
+  "version": "0.0.1",
+  "description": "Content Management services for HCMI Portal`",
   "private": true,
   "main": "index.js",
   "scripts": {
@@ -11,12 +11,13 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@arranger/server": "^0.3.57",
     "babel-polyfill": "^6.26.0",
     "dotenv": "^5.0.1",
     "elasticsearch": "^14.0.0",
     "express-inspector": "^0.3.1",
-    "fs": "^0.0.1-security"
+    "fs": "^0.0.1-security",
+    "googleapis": "27",
+    "mongoose": "^5.1.3"
   },
   "devDependencies": {
     "@babel/cli": "^7.0.0-beta.35",
